@@ -1,22 +1,49 @@
-import { Module } from '@nestjs/common';
+import { DynamicModule, Module } from '@nestjs/common';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 
-import {
-	AuthUserExtractorStrategy,
-	ExtractUserFromMsPrincipleHeader,
-} from './auth-user-extractor-strategies/auth-user-extractor.strategy';
+import { VerifyAADB2CJWTStrategy } from './auth-strategies/verify-aadb2c-jwt.strategy';
+import { VerifyAuthUserStrategy } from './auth-strategies/verify-auth-user.strategy';
+import { VerifyDevBearerStrategy } from './auth-strategies/verify-dev-bearer.strategy';
 import { AuthInterceptor } from './interceptors/auth.interceptor';
 
 @Module({
 	providers: [
 		{
-			provide: AuthUserExtractorStrategy,
-			useClass: ExtractUserFromMsPrincipleHeader,
-		},
-		{
-			provide: APP_INTERCEPTOR,
-			useClass: AuthInterceptor,
+			provide: VerifyAuthUserStrategy,
+			useClass: VerifyAADB2CJWTStrategy,
 		},
 	],
 })
-export class AuthModule {}
+class AADB2CAuthModule {}
+
+@Module({
+	providers: [
+		{
+			provide: VerifyAuthUserStrategy,
+			useClass: VerifyDevBearerStrategy,
+		},
+	],
+})
+class DevAuthModule {}
+
+const AUTH_MODULES = Object.freeze({
+	dev: DevAuthModule,
+	aadb2c: AADB2CAuthModule,
+});
+
+@Module({})
+export class AuthModule {
+	static forRoot(authProvider: keyof typeof AUTH_MODULES): DynamicModule {
+		return {
+			module: AuthModule,
+			imports: [AUTH_MODULES[authProvider]],
+			exports: [VerifyAuthUserStrategy],
+			providers: [
+				{
+					provide: APP_INTERCEPTOR,
+					useClass: AuthInterceptor,
+				},
+			],
+		};
+	}
+}
