@@ -1,4 +1,3 @@
-export * from './lib/sentry-observability.module';
-export * from './lib/dev-observability.module';
+export * from './lib/observability.module';
 export * from './lib/decorators/trace.decorator';
 export * from './lib/services/kordis-logger.interface';
