@@ -1,7 +1,6 @@
-import { Logger, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 
 import { KORDIS_LOGGER_SERVICE } from './services/kordis-logger-service.interface';
-import { KordisLoggerImpl } from './services/kordis.logger';
 import { PinoLogger } from './services/pino-logger.service';
 
 @Module({
@@ -9,10 +8,6 @@ import { PinoLogger } from './services/pino-logger.service';
 		{
 			provide: KORDIS_LOGGER_SERVICE,
 			useValue: new PinoLogger(true),
-		},
-		{
-			provide: Logger,
-			useClass: KordisLoggerImpl,
 		},
 	],
 })
